LPD-77031 Extract getFieldKey to FragmentMappingFieldUtil
