+++ modules/apps/headless/headless-admin-site/headless-admin-site-impl/src/main/java/com/liferay/headless/admin/site/internal/dto/v1_0/util/FragmentMappingFieldUtil.java
@@ -1,0 +1,38 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2026 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.headless.admin.site.internal.dto.v1_0.util;
+
+import com.liferay.portal.kernel.json.JSONObject;
+import com.liferay.portal.kernel.util.Validator;
+
+/**
+ * @author Rubén Pulido
+ */
+public class FragmentMappingFieldUtil {
+
+	public static String getFieldKey(JSONObject jsonObject) {
+		String collectionFieldId = jsonObject.getString("collectionFieldId");
+
+		if (Validator.isNotNull(collectionFieldId)) {
+			return collectionFieldId;
+		}
+
+		String fieldId = jsonObject.getString("fieldId");
+
+		if (Validator.isNotNull(fieldId)) {
+			return fieldId;
+		}
+
+		String mappedField = jsonObject.getString("mappedField");
+
+		if (Validator.isNotNull(mappedField)) {
+			return mappedField;
+		}
+
+		return null;
+	}
+
+}
